# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Engineering Document Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List all organizations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationList"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    put:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update an organization
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete an organization
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Projects
...
  /projects/{projectId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents in a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: folderId
          in: query
          schema:
            type: integer
        - name: discipline
          in: query
          schema:
            type: string
        - name: documentType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /projects/{projectId}/documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    put:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/documents/{id}/revisions:
    get:
      operationId: listDocumentRevisions
      tags: [document-revisions]
      summary: List all revisions of a document
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revisions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRevisionList"
  /projects/{projectId}/documents/{id}/submit-review:
    post:
      operationId: submitDocumentForReview
      tags: [documents]
      summary: Submit a document for review
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewRequest"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # Document Folders
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [client, consultant, contractor, subcontractor]
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        userCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - createdAt
    OrganizationList:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        total:
          type: integer
      required:
        - organizations
        - total
    CreateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [client, consultant, contractor, subcontractor]
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
      required:
        - name
        - type

    # Project
...
    Document:
      type: object
      properties:
        id:
          type: integer
        documentNumber:
          type: string
        title:
          type: string
        documentType:
          type: string
        discipline:
          type: string
        revision:
          type: string
        status:
          type: string
          enum: [draft, under_review, approved, issued, superseded, void]
        description:
          type: string
        folderId:
          type: integer
        folderName:
          type: string
        projectId:
          type: integer
        createdById:
          type: integer
        createdByName:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        metadata:
          type: object
          additionalProperties: true
        workflowStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - documentNumber
        - title
        - revision
        - status
        - projectId
        - createdAt
        - updatedAt
    DocumentList:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
      required:
        - documents
        - total
    CreateDocumentRequest:
      type: object
      properties:
        documentNumber:
          type: string
        title:
          type: string
        documentType:
          type: string
        discipline:
          type: string
        revision:
          type: string
        status:
          type: string
          enum: [draft, under_review, approved, issued, superseded, void]
        description:
          type: string
        folderId:
          type: integer
        fileUrl:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        metadata:
          type: object
          additionalProperties: true
      required:
        - documentNumber
        - title
        - revision
    UpdateDocumentRequest:
      type: object
      properties:
        title:
          type: string
        documentType:
          type: string
        discipline:
          type: string
        revision:
          type: string
        status:
          type: string
          enum: [draft, under_review, approved, issued, superseded, void]
        description:
          type: string
        folderId:
          type: integer
        fileUrl:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        metadata:
          type: object
          additionalProperties: true
    SubmitReviewRequest:
      type: object
      properties:
        reviewerIds:
          type: array
          items:
            type: integer
        comment:
          type: string
      required:
        - reviewerIds
    DocumentRevision:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        revision:
          type: string
        status:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        comment:
          type: string
        createdById:
          type: integer
        createdByName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - documentId
        - revision
        - status
        - createdAt
    DocumentRevisionList:
      type: object
      properties:
        revisions:
          type: array
          items:
            $ref: "#/components/schemas/DocumentRevision"
        total:
          type: integer
      required:
        - revisions
        - total

    # Folder
...